+++ src/main/java/com/fooddelivery/Model/Merchants.java
@@ -16,7 +16,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 public class Merchants {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private long merID;
+	private int merID;
 	
 	private String merName;
 	
@@ -48,11 +48,11 @@ public class Merchants {
 		this.cookingTime = cookingTime;
 	}
 
-	public long getMerID() {
+	public int getMerID() {
 		return merID;
 	}
 
-	public void setMerID(long merID) {
+	public void setMerID(int merID) {
 		this.merID = merID;
 	}
